% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_osm_roads.R
\name{city_osm_roads}
\alias{city_osm_roads}
\title{Extract OSM Roads for City}
\usage{
city_osm_roads(city_border, road_tags, crsLONGLAT)
}
\arguments{
\item{city_border}{A spatial object representing the boundary of the city.}

\item{road_tags}{A character vector specifying the types of roads to retrieve from OSM.}

\item{crsLONGLAT}{A character string representing the coordinate reference system (CRS) for the city boundary.}
}
\value{
A spatial object containing the OSM roads data for the specified city.
}
\description{
This function retrieves OpenStreetMap (OSM) road data for a specified city boundary and plots the roads.
}
\details{
This function retrieves road data from OpenStreetMap (OSM) for a specified city boundary and plots
the roads using ggplot2. The road types to be retrieved can be specified using the road_tags parameter.
}
\examples{
road_tags <- c("motorway", "trunk", "primary", "secondary",
"tertiary", "motorway_link", "trunk_link",
"primary_link", "secondary_link", "tertiary_link")

city_roads <- city_osm_roads(city_border_osm, road_tags, crsLONGLAT)

}
