% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{create_map}
\alias{create_map}
\title{Create Map}
\usage{
create_map(raster_df, roads_cropped, title, caption)
}
\arguments{
\item{raster_df}{A dataframe containing the values of the raster dataset.}

\item{roads_cropped}{A spatial object containing the cropped roads dataset}

\item{title}{A character string specifying the title of the map.}

\item{caption}{A character string providing additional information or caption for the map.}
}
\value{
A ggplot2 object representing the created map.
}
\description{
This function creates a map by overlaying a raster dataset and cropped roads on a ggplot2 plot.
}
\details{
This function creates a map by overlaying a raster dataset, which represents built-up areas, and cropped roads
dataset on a ggplot2 plot. The map is customized with specified title, caption, and styling parameters. Please use zoom on Plots window
to observe full size of the map.
}
\examples{
city_map_buffer <- create_map(city_df_buffer, city_roads_inside_buffer, "Istanbul Urban Expansion","Created by <name>")
city_map_boundaries <- create_map(city_df_boundaries, city_roads_inside_boundaries, "Istanbul Urban Expansion","Created by <name>")

#ggsave(
 # filename = "built_up.png",
 # width = 6, height = 6, dpi = 600,
 # device = "png", city_map_boundaries #city_map_buffer
# )
}
