% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_city_roads.R
\name{crop_city_roads}
\alias{crop_city_roads}
\title{Crop City Roads}
\usage{
crop_city_roads(road_tags, cropped_method, crsLONGLAT)
}
\arguments{
\item{road_tags}{A character vector specifying the types of roads to include.}

\item{cropped_method}{An object representing the method used for cropping the roads.}

\item{crsLONGLAT}{The coordinate reference system (CRS) for the output roads.}
}
\value{
An sf object representing the cropped city roads.
}
\description{
This function crops city roads based on specified road tags and a cropping method.
}
\examples{
road_tags <- c("motorway", "trunk", "primary", "secondary",
"tertiary", "motorway_link", "trunk_link",
"primary_link", "secondary_link", "tertiary_link")

city_roads_inside_buffer <- crop_city_roads(road_tags,crop_raster_buffer$buffer_polygon, crsLONGLAT)
city_roads_inside_boundaries <- crop_city_roads(road_tags,city_border_osm,crsLONGLAT)
}
