% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_raster_by_boundaries.R
\name{crop_raster_by_boundaries}
\alias{crop_raster_by_boundaries}
\title{Crop Raster by Boundaries}
\usage{
crop_raster_by_boundaries(raster_data, polygon_data)
}
\arguments{
\item{raster_data}{A raster object representing the raster dataset to be cropped.}

\item{polygon_data}{A vector or object containing the polygon boundaries.}
}
\value{
A cropped raster object with the specified boundaries.
}
\description{
This function crops a raster data using the boundaries defined by a polygon.
}
\examples{
crop_raster_by_boundaries<- crop_raster_by_boundaries(builtup_raster_data, city_border_osm)
}
