% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_raster_by_buffer.R
\name{crop_raster_by_buffer}
\alias{crop_raster_by_buffer}
\title{Crop Raster by Buffer}
\usage{
crop_raster_by_buffer(
  city_border,
  buffer_size_km,
  buffer_shape,
  builtup_raster_data
)
}
\arguments{
\item{city_border}{A spatial object representing the boundary of the city.}

\item{buffer_size_km}{The size of the buffer zone in kilometers.}

\item{buffer_shape}{The shape of the buffer zone, either "circle" or "square".}

\item{builtup_raster_data}{A raster object representing the built-up data.}
}
\value{
A list containing the buffer polygon and the cropped raster data.($buffer_polygon and $raster_masked)
}
\description{
This function crops a raster dataset using a buffer zone around a city boundary.
}
\details{
This function allows you to crop a raster dataset using a buffer zone around a city boundary.
The buffer can be either a circle or a square shape.
}
\examples{
buffer_size <- 30
buffer_shape <- "square"
crop_raster_by_buffer <- crop_raster_by_buffer(city_border_osm, buffer_size, buffer_shape, builtup_raster_data)
}
