% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_builtup_data.R
\name{fetch_builtup_data}
\alias{fetch_builtup_data}
\title{FETCH BUILD-UP DATA}
\usage{
fetch_builtup_data(url, lat_lon_list, crsLONGLAT)
}
\arguments{
\item{url}{A character string specifying the URL where the raster data is located.}

\item{lat_lon_list}{A character vector specifying tile-s for the locations of interest.}

\item{crsLONGLAT}{A character string representing the coordinate reference system (CRS)}
}
\value{
A single or merged raster tile layer containing the built-up data for the specified locations.
}
\description{
This function fetches built-up data from a given URL for specific locations
defined by Tile-s. (check here for your tile-s using QGIS https://glad.umd.edu/users/Potapov/GLCLUC2020/10d_tiles.zip).
}
\examples{
url <- "https://glad.umd.edu/users/Potapov/GLCLUC2020/Built-up_change_2000_2020/"
lat_lon_list <- c("50N_020E")
crsLONGLAT <- "+proj=longlat +ellps=WGS84 +towgs84=0,0,0,0,0,0,0 +no_defs"

builtup_raster_data <- fetch_builtup_data(url, lat_lon_list, crsLONGLAT)
}
