% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_city_boundaries.R
\name{get_city_boundaries}
\alias{get_city_boundaries}
\title{Get City Boundaries from OSM data}
\usage{
get_city_boundaries(city_name, raster_data, crsLONGLAT)
}
\arguments{
\item{city_name}{A character string specifying the name of the city.}

\item{raster_data}{A raster object containing the data to be plotted.}

\item{crsLONGLAT}{A character string representing the coordinate reference system CRS for the city boundaries.}
}
\value{
A sf polygon object representing the boundaries of the specified city.
}
\description{
This function retrieves the boundaries of a specified city and plots them on top of raster data.
}
\details{
This function uses the OpenStreetMap OSM API to retrieve the bounding box of the specified city,
converts it into a spatial polygon object, and plots it on top of the provided raster (buildup_raster_data) data.
}
\examples{
city <- "Istanbul"

city_border_osm <- get_city_boundaries(city, builtup_raster_data, crsLONGLAT)
}
