% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbin_map.R
\name{hexbin_map}
\alias{hexbin_map}
\title{Hexbin Map}
\usage{
hexbin_map(city_df, bins, title)
}
\arguments{
\item{city_df}{A dataframe containing city data, including x and y coordinates.}

\item{bins}{An integer specifying the number of hexagonal bins for the density map.}

\item{title}{A character string specifying the title of the map.}
}
\value{
A ggplot2 object representing the hexbin density map.
}
\description{
This function creates a hexbin density map based on a dataframe containing new builtup data.
}
\details{
This function creates a hexbin density map based on the provided city data dataframe.
It filters the dataframe to include only areas with a value of 1 (indicating new buildings), and then plots
the density of these new buildings using hexagonal bins. The map is customized with the specified title and styling parameters.
}
\examples{
hex_map_polygon <- hexbin_map(city_df_boundaries,15,"Building Desinty, Pixel-Based Analysis")
hex_map_buffer <- hexbin_map(city_df_buffer,15,"Building Desinty, Pixel-Based Analysis")
}
