% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_to_df.R
\name{raster_to_df}
\alias{raster_to_df}
\title{Convert Raster to Dataframe}
\usage{
raster_to_df(raster_data)
}
\arguments{
\item{raster_data}{}
}
\value{
A dataframe containing the values of the raster (x,y,values and cat columns).
}
\description{
This function converts a raster (cropped raster by boundaries or buffer) data to a dataframe.
}
\examples{

city_df_boundaries <- raster_to_df(crop_raster_boundaries)
city_df_buffer <- raster_to_df(crop_raster_buffer$raster_masked)
}
