% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UrbanExpansionUI.R
\name{UrbanExpansionUI}
\alias{UrbanExpansionUI}
\title{Urban Expansion Analysis Dashboard}
\usage{
UrbanExpansionUI(
  city_map_buffer,
  city_map_boundaries,
  city_df_buffer,
  city_df_boundaries
)
}
\arguments{
\item{city_map_buffer}{A map object representing the city map with buffer.}

\item{city_map_boundaries}{A map object representing the city map with boundaries.}

\item{city_df_buffer}{A data frame containing data related to the city map with buffer.}

\item{city_df_boundaries}{A data frame containing data related to the city map with boundaries.}
}
\value{
A Shiny dashboard for urban expansion analysis.
}
\description{
This function generates a Shiny dashboard for visualizing urban expansion analysis.
}
\examples{
UrbanExpansionUI(city_map_buffer,city_map_boundaries,city_df_buffer,city_df_boundaries)
}
