% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_city_roads.R
\name{crop_city_roads}
\alias{crop_city_roads}
\title{Crop City Roads}
\usage{
crop_city_roads(city_roads, cropped_method)
}
\arguments{
\item{city_roads}{A spatial object containing the city roads dataset.}

\item{cropped_method}{A spatial object representing the boundary or area to crop the city roads dataset.}
}
\value{
A spatial object containing the cropped city roads dataset.
}
\description{
This function crops the specified city roads dataset based on the provided boundary, buffer or spatial object.
}
\details{
This function intersects the city roads dataset with the specified boundary or area to crop the roads within that boundary.
}
\examples{
city_roads_inside_buffer <- crop_city_roads(city_roads,crop_raster_by_buffer$buffer_polygon)
city_roads_inside_boundaries <- crop_city_roads(city_roads,city_border_osm)
}
